Fix deadlock in record.Addr()
Goroutine A is populating r.next until r.stop has a value
Goroutine B is pulling from r.next
Goroutine C sends a value to r.stop if the record is expired

The select stmt in A recieves the stop value and exits, goroutine B is
stuck waiting for a channel that will never be populated.

This patch removes the stop channel, instead the next channel will just
be closed when the record is expired.